First public release of the RAVEnOUT perl script
OK - am not proud. It is a bit of a hack, but hey it basically works! ;-) Feel free to make changes/update/clean up.

Currently this does not [yet] replace the V1 script, but it's definitely close. 

Issues at moment - sometimes no response is received from the meter in time, in which case the script closes the file so no data is captured and the script dies. The script may be run again manually, but my next task is to put some smarts into the script to test the file and if it is not valid then to run again. At the moment I have put in a small delay, but this is obviously not a permanent fix 

To be done - I want to resolve the above issue, then will now work on option to send the data to remote services such as pvoutput and xively